deliver-assist/requirements.txt
google-genai>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
